+++ internal/app/wwctl/container/copy/main.go
@@ -4,7 +4,6 @@ import (
 	"fmt"
 
 	"github.com/spf13/cobra"
-	"github.com/warewulf/warewulf/internal/pkg/api/routes/wwapiv1"
 	"github.com/warewulf/warewulf/internal/pkg/container"
 	"github.com/warewulf/warewulf/internal/pkg/wwlog"
 )
@@ -15,34 +14,18 @@ func CobraRunE(cmd *cobra.Command, args []string) (err error) {
 		wwlog.Warn("copy only requires 2 arguments but you provided %d arguments. Hence, they will be ignored.", len(args))
 	}
 
-	cdp := &wwapiv1.ContainerCopyParameter{
-		ContainerSource:      args[0],
-		ContainerDestination: args[1],
-	}
-
-	if !container.DoesSourceExist(cdp.ContainerSource) {
-		wwlog.Error("Container's source doesn't exists: %s", cdp.ContainerSource)
-		return
-	}
-
-	if !container.ValidName(cdp.ContainerDestination) {
-		wwlog.Error("Container name contains illegal characters : %s", cdp.ContainerDestination)
-		return
-	}
-
-	if container.DoesSourceExist(cdp.ContainerDestination) {
-		wwlog.Error("An other container with name: %s already exists in sources.", cdp.ContainerDestination)
-		return
-	}
-
-	err = container.Duplicate(cdp.ContainerSource, cdp.ContainerDestination)
+	err = container.Copy(&container.CopyParameter{
+		Name:        args[0],
+		Destination: args[1],
+		ForceBuild:  true,
+	})
 	if err != nil {
 		err = fmt.Errorf("could not duplicate image: %s", err.Error())
 		wwlog.Error(err.Error())
 		return
 	}
 
-	wwlog.Info("Container %s successfully duplicated as %s", cdp.ContainerSource, cdp.ContainerDestination)
+	wwlog.Info("Container %s successfully duplicated as %s", args[0], args[1])
 	return
 
 }