+++ internal/pkg/container/build.go
@@ -1,15 +1,56 @@
 package container
 
 import (
+	"fmt"
 	"path"
 
 	"github.com/pkg/errors"
-
 	"github.com/warewulf/warewulf/internal/pkg/util"
 	"github.com/warewulf/warewulf/internal/pkg/wwlog"
+	"golang.org/x/sync/errgroup"
 )
 
-func Build(name string, buildForce bool) error {
+type BuildParameter struct {
+	Names   []string
+	Force   bool
+	Default bool
+}
+
+func Build(param *BuildParameter) error {
+	if len(param.Names) == 0 {
+		return fmt.Errorf("build names array is empty")
+	}
+
+	if param.Default && len(param.Names) != 1 {
+		return fmt.Errorf("can only set default for one container")
+	}
+
+	g := new(errgroup.Group)
+
+	for _, name := range param.Names {
+		if !ValidSource(name) {
+			return fmt.Errorf("build source %s does not exist", name)
+		}
+
+		n := name
+		g.Go(func() error {
+			err := build(n, param.Force)
+			return err
+		})
+	}
+
+	if err := g.Wait(); err != nil {
+		return fmt.Errorf("error occurs during build, err: %s", err)
+	}
+
+	if param.Default {
+		return SetProfileDefaultContainer(param.Names[0])
+	}
+
+	return nil
+}
+
+func build(name string, buildForce bool) error {
 
 	rootfsPath := RootFsDir(name)
 	imagePath := ImageFile(name)